Remove ConsoleLoggerProvider ctor overloads
These don't have much usage and their addition could create problems for
DI Providers.
