+++ src/libraries/Microsoft.Extensions.Logging.Console/src/ConsoleLoggerExtensions.Obsolete.cs
@@ -5,6 +5,8 @@ using System;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging.Configuration;
 using Microsoft.Extensions.Logging.Console;
+using Microsoft.Extensions.Options;
+using Microsoft.Extensions.Primitives;
 
 namespace Microsoft.Extensions.Logging
 {
@@ -20,7 +22,7 @@ namespace Microsoft.Extensions.Logging
         [Obsolete("This method is retained only for compatibility. The recommended alternative is AddConsole(this ILoggingBuilder builder).", error: true)]
         public static Logging.ILoggerFactory AddConsole(this Logging.ILoggerFactory factory, Console.IConsoleLoggerSettings settings)
         {
-            factory.AddProvider(new ConsoleLoggerProvider(settings));
+            factory.AddProvider(new ConsoleLoggerProvider(ConsoleLoggerSettingsAdapter.GetOptionsMonitor(settings)));
             return factory;
         }
 
@@ -48,7 +50,7 @@ namespace Microsoft.Extensions.Logging
         [Obsolete("This method is retained only for compatibility. The recommended alternative is AddConsole(this ILoggingBuilder builder).", error: true)]
         public static Logging.ILoggerFactory AddConsole(this Logging.ILoggerFactory factory, System.Func<string, Logging.LogLevel, bool> filter, bool includeScopes)
         {
-            factory.AddProvider(new ConsoleLoggerProvider(filter, includeScopes));
+            factory.AddConsole(new ConsoleLoggerSettings() { IncludeScopes = includeScopes });
             return factory;
         }
 
@@ -63,6 +65,53 @@ namespace Microsoft.Extensions.Logging
         public static Logging.ILoggerFactory AddConsole(this Logging.ILoggerFactory factory)
         {
             return factory.AddConsole(includeScopes: false);
+        }
+
+        [Obsolete]
+        private sealed class ConsoleLoggerSettingsAdapter : IConfigureOptions<ConsoleLoggerOptions>, IOptionsChangeTokenSource<ConsoleLoggerOptions>
+        {
+            private IConsoleLoggerSettings _settings;
+            private ConsoleLoggerSettingsAdapter(IConsoleLoggerSettings settings)
+            {
+                _settings = settings;
+            }
+
+            IChangeToken IOptionsChangeTokenSource<ConsoleLoggerOptions>.GetChangeToken() => _settings.ChangeToken ?? NullChangeToken.Instance;
+
+            string IOptionsChangeTokenSource<ConsoleLoggerOptions>.Name => Microsoft.Extensions.Options.Options.DefaultName;
+
+            void IConfigureOptions<ConsoleLoggerOptions>.Configure(ConsoleLoggerOptions options)
+            {
+                options.IncludeScopes = _settings.IncludeScopes;
+                if (_settings is ConfigurationConsoleLoggerSettings configSettings)
+                {
+                    options.Configure(configSettings._configuration);
+                }
+                else if (_settings is ConsoleLoggerSettings consoleSettings)
+                {
+                    options.DisableColors = consoleSettings.DisableColors;
+                }
+            }
+
+            internal static OptionsMonitor<ConsoleLoggerOptions> GetOptionsMonitor(IConsoleLoggerSettings settings)
+            {
+                ConsoleLoggerSettingsAdapter adapter = new(settings);
+                OptionsFactory<ConsoleLoggerOptions> factory = new( new IConfigureOptions<ConsoleLoggerOptions>[] { adapter }, Array.Empty<IPostConfigureOptions<ConsoleLoggerOptions>>());
+                IOptionsChangeTokenSource<ConsoleLoggerOptions>[] sources = new IOptionsChangeTokenSource<ConsoleLoggerOptions>[] { adapter };
+                OptionsCache<ConsoleLoggerOptions> cache = new();
+
+                return new OptionsMonitor<ConsoleLoggerOptions>(factory, sources, cache);
+            }
+        }
+
+        private sealed class NullChangeToken : IChangeToken, IDisposable
+        {
+            internal static NullChangeToken Instance { get; } = new NullChangeToken();
+            private NullChangeToken() { }
+            public bool HasChanged => false;
+            public bool ActiveChangeCallbacks => false;
+            public IDisposable RegisterChangeCallback(Action<object?> callback, object? state) => this;
+            public void Dispose() { }
         }
     }
 }
